feat(packages/sui-mono): use preset
